🔀 Merge Develop (Temporary profile change for db creation) (#22)

+++ src/main/kotlin/com/wafflestudio/team2/jisik2n/core/answer/service/AnswerService.kt
@@ -1,8 +1,12 @@
 package com.wafflestudio.team2.jisik2n.core.answer.service
 
+import com.wafflestudio.team2.jisik2n.common.Jisik2n400
+import com.wafflestudio.team2.jisik2n.common.Jisik2n403
+import com.wafflestudio.team2.jisik2n.common.Jisik2n404
 import com.wafflestudio.team2.jisik2n.core.answer.database.AnswerEntity
 import com.wafflestudio.team2.jisik2n.core.answer.database.AnswerRepository
 import com.wafflestudio.team2.jisik2n.core.answer.dto.AnswerRequest
+import com.wafflestudio.team2.jisik2n.core.answer.dto.AnswerResponse
 import com.wafflestudio.team2.jisik2n.core.photo.database.PhotoEntity
 import com.wafflestudio.team2.jisik2n.core.photo.database.PhotoRepository
 import com.wafflestudio.team2.jisik2n.core.question.database.QuestionRepository
@@ -10,9 +14,13 @@ import com.wafflestudio.team2.jisik2n.core.user.database.UserEntity
 import com.wafflestudio.team2.jisik2n.core.user.database.UserRepository
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
+import java.time.LocalDateTime
+import java.time.ZoneId
 import javax.transaction.Transactional
 
 interface AnswerService {
+    fun getAnswersOfQuestion(questionId: Long): List<AnswerResponse>
+
     fun createAnswer(
         loginUser: UserEntity,
         questionId: Long,
@@ -25,6 +33,8 @@ interface AnswerService {
         answerRequest: AnswerRequest
     )
 
+    fun toggleSelectAnswer(loginUser: UserEntity, answerId: Long, toSelect: Boolean)
+
     fun removeAnswer(loginUser: UserEntity, answerId: Long)
 }
 
@@ -35,6 +45,17 @@ class AnswerServiceImpl(
     private val photoRepository: PhotoRepository,
     private val userRepository: UserRepository,
 ) : AnswerService {
+    override fun getAnswersOfQuestion(questionId: Long): List<AnswerResponse> {
+        // Get target question
+        val question = questionRepository.findByIdOrNull(questionId)
+            ?: throw Jisik2n404("${questionId}에 해당하는 질문이 없습니다.")
+
+        // TODO: Improve query
+        val answers = question.answers.sortedWith(compareBy({ !it.selected }, { it.createdAt }))
+
+        return answers.map { it.toResponse(answerRepository) }
+    }
+
     @Transactional
     override fun createAnswer(
         loginUser: UserEntity,
@@ -43,7 +64,11 @@ class AnswerServiceImpl(
     ) {
         // Get target question
         val question = questionRepository.findByIdOrNull(questionId)
-            ?: TODO("Throw 404 Exception")
+            ?: throw Jisik2n404("${questionId}에 해당하는 질문이 없습니다.")
+
+        if (loginUser.id == question.user.id) {
+            throw Jisik2n403("자신의 질문에는 답할 수 없습니다.")
+        }
 
         // Add new answer
         var newAnswer = answerRequest.let {
@@ -61,18 +86,6 @@ class AnswerServiceImpl(
             newAnswer.photos.addAll(it)
         }
         newAnswer = answerRepository.save(newAnswer)
-
-        // Add newAnswer to question
-        question.let {
-            it.answers.add(newAnswer)
-            questionRepository.save(it)
-        }
-
-        // Add newAnswer to user answers
-        loginUser.let {
-            it.answers.add(newAnswer)
-            userRepository.save(it)
-        }
     }
 
     @Transactional
@@ -82,10 +95,13 @@ class AnswerServiceImpl(
         answerRequest: AnswerRequest
     ) {
         val answer = answerRepository.findByIdOrNull(answerId)
-            ?: TODO("Throw 404 Not Found Exception")
+            ?: throw Jisik2n404("${answerId}에 해당하는 답변이 없습니다")
 
-        if (answer.user.id != loginUser.id || answer.selected) {
-            TODO("Throw 403 Forbidden Exception")
+        if (answer.user.id != loginUser.id) {
+            throw Jisik2n403("자신의 게시물만 수정할 수 있습니다.")
+        }
+        if (answer.question.close) {
+            throw Jisik2n403("마감된 질문은 수정될 수 없습니다.")
         }
 
         // Update content
@@ -113,14 +129,51 @@ class AnswerServiceImpl(
     }
 
     @Transactional
+    override fun toggleSelectAnswer(loginUser: UserEntity, answerId: Long, toSelect: Boolean) {
+        val answer = answerRepository.findByIdOrNull(answerId) // TODO: Optimize query (by fetch question-user)
+            ?: throw Jisik2n404("$answerId 답변이 존재하지 않습니다.")
+
+        if (loginUser.id != answer.question.user.id) {
+            throw Jisik2n403("질문한 사용자만 답변을 채택할 수 있습니다.")
+        }
+
+        if (answer.selected == toSelect) {
+            throw Jisik2n400("변경할 점이 없습니다.")
+        }
+
+        if (toSelect) { selectAnswer(answer) } else { unselectAnswer(answer) }
+    }
+
+    @Transactional
+    fun selectAnswer(answer: AnswerEntity) {
+        if (answer.question.close) {
+            throw Jisik2n403("마감된 질문에 답변을 채택할 수 없습니다.")
+        }
+
+        answer.selected = true
+        answer.selectedAt = LocalDateTime.now(ZoneId.of("Asia/Seoul"))
+        answer.question.close = true
+    }
+
+    @Transactional
+    fun unselectAnswer(answer: AnswerEntity) {
+        answer.selected = false
+        answer.question.close = false
+    }
+
+    @Transactional
     override fun removeAnswer(loginUser: UserEntity, answerId: Long) {
         val answer = answerRepository.findByIdOrNull(answerId)
 
         if (answer != null) {
-            if (answer.user.id != loginUser.id || answer.selected) {
-                TODO("Throw 403 Forbidden Exception")
+            if (answer.user.id != loginUser.id) {
+                throw Jisik2n403("자신의 게시물만 삭제할 수 있습니다.")
+            }
+            if (answer.question.close) {
+                throw Jisik2n403("마감된 질문은 삭제될 수 없습니다.")
             }
             answerRepository.deleteById(answerId)
+            // TODO: Remove Interactions
         }
     }
 }
